headless selenium use; attempt to fix requests socket warning; moved SSRN into tests.py
